fix: fix calculation goal and metric

+++ Monhealth.Application/Contracts/Services/IMetricsCalculator.cs
@@ -2,7 +2,7 @@
 {
     public interface IMetricsCalculator
     {
-        double CalculateBMI(double weight, double height);
+        float CalculateBMI(float weight, float height);
         float CalculateBMR(float weight, float height, int age, string gender);
         float CalculateTDEE(float bmr, float activityLevel);
         float CalculateIBW(float height, string gender);
